requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
billiard==4.2.2
black==25.9.0
blinker==1.9.0
brotli==1.2.0
celery==5.5.3
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
click-default-group==1.2.4
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==46.0.3
cssselect2==0.8.0
databases==0.9.0
dnspython==2.8.0
ecdsa==0.19.1
Faker==37.12.0
fastapi==0.119.0
fastapi-cli==0.0.13
fastapi-cloud-cli==0.3.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
humanize==4.14.0
idna==3.11
itsdangerous==2.2.0
kombu==5.5.4
lia-web==0.2.3
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mypy_extensions==1.1.0
orjson==3.11.3
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.4.0
pluggy==1.6.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.0
pydantic-extra-types==2.10.6
pydantic-settings==2.11.0
pydantic_core==2.41.1
pydyf==0.12.1
Pygments==2.19.2
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytokens==0.1.10
pytz==2025.2
PyYAML==6.0.3
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.0
rsa==4.9.1
sentry-sdk==2.42.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
sqlite-fts4==1.0.3
sqlite-utils==3.38
tabulate==0.9.0
tinycss2==1.5.1
tinyhtml5==2.0.0
tornado==6.5.2
txt==2020.11.12
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
urllib3==2.5.0
uvicorn==0.37.
